refactor(i18n): replace translator with formatMessage and useTranslator hook in components and tests - Updated ActionMenuComponent, CVCInput, SelectInput, and ModalComponent - Modified snapshot tests to reflect changes in translation keys and values.
